# src/main/resources/static/docs/openapi3.yaml
openapi: 3.0.1
info:
  title: QUESTRIP API
  description: QUESTRIP API description
  version: 1.0.0
servers:
- url: https://api.questrips.com
- url: http://localhost:8080
tags: []
paths:
  /api/v1/place:
    get:
      tags:
      - place
      summary: 장소 리스트 조회 API
      description: 장소 리스트 조회 API
      operationId: place-list-get
      parameters:
      - name: latitude
        in: query
        description: 유저 위도
        required: true
        schema:
          type: string
      - name: longitude
        in: query
        description: 유저 경도
        required: true
        schema:
          type: string
      - name: page
        in: query
        description: 요청 페이지 (default 0)
        required: true
        schema:
          type: string
      - name: size
        in: query
        description: 요청 사이즈 (default 10)
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-place893682493'
              examples:
                place-list-get:
                  value: "{\"status\":\"SUCCESS\",\"message\":\"장소 조회 성공\",\"data\"\
                    :{\"content\":[{\"id\":\"6633897aa2757d5b1998ba0d\",\"googlePlaceId\"\
                    :\"ChIJo0gMXbOlfDURSjmLcTy52qQ\",\"placeName\":\"Kkochijib\",\"\
                    primaryType\":\"japanese_restaurant\",\"formattedAddress\":\"\
                    South Korea, Seoul, Songpa-gu, 잠실본동 314-5\",\"location\":{\"latitude\"\
                    :37.5051148,\"longitude\":127.08420179999999},\"content\":{\"\
                    recommendationReason\":\"testRecommend\",\"activity\":\"testActivity\"\
                    },\"images\":[{\"sequence\":1,\"url\":\"https://questrip-reward.s3.ap-northeast-2.amazonaws.com/836d62f7-b1a7-4550-b137-f292ea5f881c.HEIC\"\
                    },{\"sequence\":2,\"url\":\"https://questrip-reward.s3.ap-northeast-2.amazonaws.com/e2f8b8cc-fd53-47da-bea5-c113f8f6393c.jpeg\"\
                    }],\"openingHours\":[\"Monday: 5:00PM-2:00AM\",\"Tuesday: 5:00PM-2:00AM\"\
                    ,\"Wednesday: 5:00PM-2:00AM\",\"Thursday: 5:00PM-2:00AM\",\"Friday:\
                    \ 5:00PM-2:00AM\",\"Saturday: 5:00PM-2:00AM\",\"Sunday: 5:00PM-2:00AM\"\
                    ],\"openNow\":\"OPEN\",\"distance\":17.47104241758488}],\"page\"\
                    :0,\"size\":10,\"numberOfElements\":1,\"hasNext\":false}}"
    post:
      tags:
      - place
      summary: 장소 등록 API
      description: 장소 등록 API
      operationId: place-create
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/api-v1-place1222034206'
            examples:
              place-create:
                value: googlePlaceId=6633897aa2757d5b1998ba0d&recommendationReason=test&activity=test
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-place-1632714624'
              examples:
                place-create:
                  value: "{\"status\":\"SUCCESS\",\"message\":\"장소 저장 성공\",\"data\"\
                    :{\"id\":\"6633897aa2757d5b1998ba0d\",\"googlePlaceId\":\"ChIJo0gMXbOlfDURSjmLcTy52qQ\"\
                    ,\"placeName\":\"Kkochijib\",\"primaryType\":\"japanese_restaurant\"\
                    ,\"formattedAddress\":\"South Korea, Seoul, Songpa-gu, 잠실본동 314-5\"\
                    ,\"location\":{\"latitude\":37.5051148,\"longitude\":127.08420179999999},\"\
                    content\":{\"recommendationReason\":\"testRecommend\",\"activity\"\
                    :\"testActivity\"},\"images\":[{\"sequence\":1,\"url\":\"https://questrip-reward.s3.ap-northeast-2.amazonaws.com/836d62f7-b1a7-4550-b137-f292ea5f881c.HEIC\"\
                    },{\"sequence\":2,\"url\":\"https://questrip-reward.s3.ap-northeast-2.amazonaws.com/e2f8b8cc-fd53-47da-bea5-c113f8f6393c.jpeg\"\
                    }],\"openingHours\":[\"Monday: 5:00PM-2:00AM\",\"Tuesday: 5:00PM-2:00AM\"\
                    ,\"Wednesday: 5:00PM-2:00AM\",\"Thursday: 5:00PM-2:00AM\",\"Friday:\
                    \ 5:00PM-2:00AM\",\"Saturday: 5:00PM-2:00AM\",\"Sunday: 5:00PM-2:00AM\"\
                    ],\"openNow\":\"OPEN\"}}"
  /api/v1/question:
    post:
      tags:
      - question
      summary: 문의 등록 API
      description: 문의 등록 API
      operationId: question-post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-question1223598169'
            examples:
              question-post:
                value: "{\"content\":\"서비스 문의드립니다.\",\"userEmail\":\"newjeans@queen.kpop\"\
                  ,\"category\":\"SERVICE_QUESTION\"}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-user-checkEmail1876876947'
              examples:
                question-post:
                  value: "{\"status\":\"SUCCESS\",\"message\":\"문의 등록 완료\",\"data\"\
                    :null}"
  /api/v1/place/reverseGeocode:
    get:
      tags:
      - place
      summary: 리버스 지오코딩 API
      description: 리버스 지오코딩 API
      operationId: place-geocode
      parameters:
      - name: latitude
        in: query
        description: 유저 위도
        required: true
        schema:
          type: string
      - name: longitude
        in: query
        description: 유저 경도
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-place-reverseGeocode-2100626994'
              examples:
                place-geocode:
                  value: "{\"status\":\"SUCCESS\",\"message\":\"주소 변환 성공\",\"data\"\
                    :{\"formattedAddress\":\"226-19 Jamsil-dong, Songpa District,\
                    \ Seoul, South Korea\"}}"
  /api/v1/place/{placeId}:
    get:
      tags:
      - place
      summary: 장소 조회 API
      description: 장소 조회 API
      operationId: place-get
      parameters:
      - name: placeId
        in: path
        description: 플레이스 아이디(구글 x)
        required: true
        schema:
          type: string
      - name: latitude
        in: query
        description: 유저 위도
        required: true
        schema:
          type: string
      - name: longitude
        in: query
        description: 유저 경도
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-place-placeId-2120279981'
              examples:
                place-get:
                  value: "{\"status\":\"SUCCESS\",\"message\":\"장소 조회 성공\",\"data\"\
                    :{\"place\":{\"id\":\"6633897aa2757d5b1998ba0d\",\"googlePlaceId\"\
                    :\"ChIJo0gMXbOlfDURSjmLcTy52qQ\",\"placeName\":\"Kkochijib\",\"\
                    primaryType\":\"japanese_restaurant\",\"formattedAddress\":\"\
                    South Korea, Seoul, Songpa-gu, 잠실본동 314-5\",\"location\":{\"latitude\"\
                    :37.5051148,\"longitude\":127.08420179999999},\"content\":{\"\
                    recommendationReason\":\"testRecommend\",\"activity\":\"testActivity\"\
                    },\"images\":[{\"sequence\":1,\"url\":\"https://questrip-reward.s3.ap-northeast-2.amazonaws.com/836d62f7-b1a7-4550-b137-f292ea5f881c.HEIC\"\
                    },{\"sequence\":2,\"url\":\"https://questrip-reward.s3.ap-northeast-2.amazonaws.com/e2f8b8cc-fd53-47da-bea5-c113f8f6393c.jpeg\"\
                    }],\"openingHours\":[\"Monday: 5:00PM-2:00AM\",\"Tuesday: 5:00PM-2:00AM\"\
                    ,\"Wednesday: 5:00PM-2:00AM\",\"Thursday: 5:00PM-2:00AM\",\"Friday:\
                    \ 5:00PM-2:00AM\",\"Saturday: 5:00PM-2:00AM\",\"Sunday: 5:00PM-2:00AM\"\
                    ],\"openNow\":\"OPEN\"},\"directionSummary\":{\"distance\":\"\
                    27.1 km\",\"duration\":\"1 hour 16 mins\"}}}"
  /api/v1/user/checkEmail:
    post:
      tags:
      - user
      summary: 이메일 중복검사 API
      description: 이메일 중복검사 API
      operationId: user-checkEmail
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-user-checkEmail1625218753'
            examples:
              user-checkEmail:
                value: "{\"email\":\"harok@questrips.com\"}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-user-checkEmail1876876947'
              examples:
                user-checkEmail:
                  value: "{\"status\":\"SUCCESS\",\"message\":\"이메일 사용 가능\",\"data\"\
                    :null}"
  /api/v1/user/join:
    post:
      tags:
      - user
      summary: 회원가입 API
      description: 회원가입 API
      operationId: user-create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-user-join-167844841'
            examples:
              user-create:
                value: "{\"username\":\"harok\",\"email\":\"harok@questrips.com\"\
                  ,\"password\":\"harok123\"}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-user-login1984296762'
              examples:
                user-create:
                  value: "{\"status\":\"SUCCESS\",\"message\":\"회원가입 완료\",\"data\"\
                    :{\"user\":{\"id\":1,\"username\":\"harok\",\"email\":\"harok@questrips.com\"\
                    ,\"role\":\"USER\"},\"accessToken\":\"accessToken\"}}"
  /api/v1/user/login:
    post:
      tags:
      - user
      summary: 로그인 API
      description: 로그인 API
      operationId: user-login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api-v1-user-login1120333644'
            examples:
              user-login:
                value: "{\"email\":\"harok@questrips.com\",\"password\":\"harok123\"\
                  }"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-v1-user-login1984296762'
              examples:
                user-login:
                  value: "{\"status\":\"SUCCESS\",\"message\":\"로그인 완료\",\"data\"\
                    :{\"user\":{\"id\":1,\"username\":\"harok\",\"email\":\"harok@questrips.com\"\
                    ,\"role\":\"USER\"},\"accessToken\":\"accessToken\"}}"
  /test:
    get:
      tags:
      - test
      operationId: test
      parameters:
      - name: Authorization
        in: header
        description: token
        required: true
        schema:
          type: string
        example: Access Token
      responses:
        "200":
          description: "200"
          content:
            text/plain;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/test486549215'
              examples:
                test:
                  value: harok@questrips.com
components:
  schemas:
    api-v1-place-reverseGeocode-2100626994:
      required:
      - data
      - message
      - status
      type: object
      properties:
        data:
          required:
          - formattedAddress
          type: object
          properties:
            formattedAddress:
              type: string
              description: 주소
          description: 데이터
        message:
          type: string
          description: 메시지
        status:
          type: string
          description: 응답 상태
    api-v1-question1223598169:
      required:
      - category
      - content
      - userEmail
      type: object
      properties:
        userEmail:
          type: string
          description: 유저 메일
        category:
          type: string
          description: 문의 유형
        content:
          type: string
          description: 내용
    api-v1-place1222034206:
      required:
      - activity
      - googlePlaceId
      - recommendationReason
      type: object
      properties:
        activity:
          type: string
          description: 추천 활동
        googlePlaceId:
          type: string
          description: 구글 플레이스 아이디
        recommendationReason:
          type: string
          description: 추천 이유
    api-v1-user-join-167844841:
      required:
      - email
      - password
      - username
      type: object
      properties:
        password:
          type: string
          description: 비밀번호
        email:
          type: string
          description: 유저 메일
        username:
          type: string
          description: 유저 이름
    api-v1-place893682493:
      required:
      - data
      - message
      - status
      type: object
      properties:
        data:
          required:
          - content
          - hasNext
          - numberOfElements
          - page
          - size
          type: object
          properties:
            numberOfElements:
              type: number
              description: 실제 데이터 개수
            size:
              type: number
              description: 요청 사이즈
            hasNext:
              type: boolean
              description: 다음 페이지 존재 여부
            page:
              type: number
              description: 요청 페이지
            content:
              type: array
              description: 데이터
              items:
                required:
                - content
                - distance
                - formattedAddress
                - googlePlaceId
                - id
                - images
                - location
                - openNow
                - openingHours
                - placeName
                - primaryType
                type: object
                properties:
                  images:
                    type: array
                    description: 장소 이미지
                    items:
                      required:
                      - sequence
                      - url
                      type: object
                      properties:
                        sequence:
                          type: number
                          description: 이미지 순서
                        url:
                          type: string
                          description: 이미지 url
                  distance:
                    type: number
                    description: 현재 위치에서 장소까지 거리(km)
                  formattedAddress:
                    type: string
                    description: 주소
                  googlePlaceId:
                    type: string
                    description: 구글 장소 아이디
                  primaryType:
                    type: string
                    description: 카테고리
                  openNow:
                    type: string
                    description: "현재 영업 여부 [OPEN, CLOSE, UNKNOWN]"
                  openingHours:
                    type: array
                    description: 영업시간
                    items:
                      oneOf:
                      - type: object
                      - type: boolean
                      - type: string
                      - type: number
                  location:
                    required:
                    - latitude
                    - longitude
                    type: object
                    properties:
                      latitude:
                        type: number
                        description: 위도
                      longitude:
                        type: number
                        description: 경도
                    description: 위경도
                  id:
                    type: string
                    description: 장소 아이디
                  placeName:
                    type: string
                    description: 장소 이름
                  content:
                    required:
                    - activity
                    - recommendationReason
                    type: object
                    properties:
                      activity:
                        type: string
                        description: 추천 활
                      recommendationReason:
                        type: string
                        description: 추천 이유
                    description: 장소 설명
          description: 데이터
        message:
          type: string
          description: 메시지
        status:
          type: string
          description: 응답 상태
    api-v1-user-login1120333644:
      required:
      - email
      - password
      type: object
      properties:
        password:
          type: string
          description: 비밀번호
        email:
          type: string
          description: 유저 메일
    api-v1-place-placeId-2120279981:
      required:
      - data
      - message
      - status
      type: object
      properties:
        data:
          required:
          - directionSummary
          - place
          type: object
          properties:
            directionSummary:
              required:
              - distance
              - duration
              type: object
              properties:
                duration:
                  type: string
                  description: 현재 위치에서 장소까지 시간
                distance:
                  type: string
                  description: 현재 위치에서 장소까지 거리(km)
              description: 길찾기 요약 정보
            place:
              required:
              - content
              - formattedAddress
              - googlePlaceId
              - id
              - images
              - location
              - openNow
              - openingHours
              - placeName
              - primaryType
              type: object
              properties:
                images:
                  type: array
                  description: 장소 이미지
                  items:
                    required:
                    - sequence
                    - url
                    type: object
                    properties:
                      sequence:
                        type: number
                        description: 이미지 순서
                      url:
                        type: string
                        description: 이미지 url
                formattedAddress:
                  type: string
                  description: 주소
                googlePlaceId:
                  type: string
                  description: 구글 장소 아이디
                primaryType:
                  type: string
                  description: 카테고리
                openNow:
                  type: string
                  description: "현재 영업 여부 [OPEN, CLOSE, UNKNOWN]"
                openingHours:
                  type: array
                  description: 영업시간
                  items:
                    oneOf:
                    - type: object
                    - type: boolean
                    - type: string
                    - type: number
                location:
                  required:
                  - latitude
                  - longitude
                  type: object
                  properties:
                    latitude:
                      type: number
                      description: 위도
                    longitude:
                      type: number
                      description: 경도
                  description: 위경도
                id:
                  type: string
                  description: 장소 아이디
                placeName:
                  type: string
                  description: 장소 이름
                content:
                  required:
                  - activity
                  - recommendationReason
                  type: object
                  properties:
                    activity:
                      type: string
                      description: 추천 활
                    recommendationReason:
                      type: string
                      description: 추천 이유
                  description: 장소 설명
              description: 플레이스
          description: 데이터
        message:
          type: string
          description: 메시지
        status:
          type: string
          description: 응답 상태
    api-v1-user-checkEmail1876876947:
      required:
      - data
      - message
      - status
      type: object
      properties:
        message:
          type: string
          description: 메시지
        status:
          type: string
          description: 응답 상태
    api-v1-place-1632714624:
      required:
      - data
      - message
      - status
      type: object
      properties:
        data:
          required:
          - content
          - formattedAddress
          - googlePlaceId
          - id
          - images
          - location
          - openNow
          - openingHours
          - placeName
          - primaryType
          type: object
          properties:
            images:
              type: array
              description: 장소 이미지
              items:
                required:
                - sequence
                - url
                type: object
                properties:
                  sequence:
                    type: number
                    description: 이미지 순서
                  url:
                    type: string
                    description: 이미지 url
            formattedAddress:
              type: string
              description: 주소
            googlePlaceId:
              type: string
              description: 구글 장소 아이디
            primaryType:
              type: string
              description: 카테고리
            openNow:
              type: string
              description: "현재 영업 여부 [OPEN, CLOSE, UNKNOWN]"
            openingHours:
              type: array
              description: 영업시간
              items:
                oneOf:
                - type: object
                - type: boolean
                - type: string
                - type: number
            location:
              required:
              - latitude
              - longitude
              type: object
              properties:
                latitude:
                  type: number
                  description: 위도
                longitude:
                  type: number
                  description: 경도
              description: 위경도
            id:
              type: string
              description: 장소 아이디
            placeName:
              type: string
              description: 장소 이름
            content:
              required:
              - activity
              - recommendationReason
              type: object
              properties:
                activity:
                  type: string
                  description: 추천 활
                recommendationReason:
                  type: string
                  description: 추천 이유
              description: 장소 설명
          description: 데이터
        message:
          type: string
          description: 메시지
        status:
          type: string
          description: 응답 상태
    api-v1-user-login1984296762:
      required:
      - data
      - message
      - status
      type: object
      properties:
        data:
          required:
          - accessToken
          - user
          type: object
          properties:
            accessToken:
              type: string
              description: 토큰
            user:
              required:
              - email
              - id
              - role
              - username
              type: object
              properties:
                role:
                  type: string
                  description: 유저 권한
                id:
                  type: number
                  description: 유저 아이디
                email:
                  type: string
                  description: 유저 이메일
                username:
                  type: string
                  description: 유저 이름
              description: 유저
          description: 데이터
        message:
          type: string
          description: 메시지
        status:
          type: string
          description: 응답 상태
    test486549215:
      type: object
    api-v1-user-checkEmail1625218753:
      required:
      - email
      type: object
      properties:
        email:
          type: string
          description: 유저 메일